Sort recipe card lists top-to-bottom
Previously, thanks to the order in which stuff goes into the LaTeX file,
the two-column recipe lists for the cards were right-to-left then
top-to-bottom whereas what we really want is top-to-bottom and then
right-to-left.  I really don't like this solution but it's pragmatic.
I'm starting to think that moving a bunch of the logic from the LaTeX
.sty file into the Python might be a good idea...